% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{pm.colors.fractions}
\alias{pm.colors.fractions}
\title{Canonical colour scheme for the fractions sampled}
\usage{
pm.colors.fractions(...)
}
\arguments{
\item{...}{ignored}
}
\value{
a named, vector of hex colors, indexed by fraction names
}
\description{
Canonical colour scheme for the fractions sampled
}
