% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{pm.colors.genotypes}
\alias{pm.colors.genotypes}
\title{Canonical colour scheme for the Arabidopsis thaliana Geontypes}
\usage{
pm.colors.genotypes(...)
}
\arguments{
\item{...}{ignored}
}
\value{
a named, vector of hex colors, indexed by Genotype names
}
\description{
Canonical colour scheme for the Arabidopsis thaliana Geontypes
}
