% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{pm.names.fractions}
\alias{pm.names.fractions}
\title{Get the list of fractions samples acrros several experiments}
\usage{
pm.names.fractions(...)
}
\value{
a vector of the canonical names of the fractions
}
\description{
Get the list of fractions samples acrros several experiments
}
