% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{pm.names.phyla}
\alias{pm.names.phyla}
\title{Get the list of Genotypes used so far}
\usage{
pm.names.phyla(...)
}
\value{
a vector of the canonical names of the genotypes
}
\description{
Get the list of Genotypes used so far
}
